REFACTOR: extract trade dialog
